+++ examples/broadcast/components/Wallet/index.tsx
@@ -1,0 +1,62 @@
+import React, { FC, useMemo } from 'react';
+import {
+  ConnectionProvider,
+  WalletProvider,
+} from '@solana/wallet-adapter-react';
+import { WalletAdapterNetwork } from '@solana/wallet-adapter-base';
+import {
+  GlowWalletAdapter,
+  PhantomWalletAdapter,
+  SolletExtensionWalletAdapter,
+  SolletWalletAdapter,
+} from '@solana/wallet-adapter-wallets';
+import {
+  WalletModalProvider,
+  WalletMultiButton,
+} from '@solana/wallet-adapter-react-ui';
+import { clusterApiUrl } from '@solana/web3.js';
+import { PhantomIcon } from '../Icon';
+
+// Default styles that can be overridden by your app
+require('@solana/wallet-adapter-react-ui/styles.css');
+
+export const Wallet = () => {
+  return (
+    <>
+      <WalletModalProvider>
+        <WalletMultiButton
+          className="dialect-wallet-button"
+          startIcon={<PhantomIcon />}
+        />
+      </WalletModalProvider>
+    </>
+  );
+};
+
+export const WalletContext: FC = (props) => {
+  // Can be set to 'devnet', 'testnet', or 'mainnet-beta'
+  const network = WalletAdapterNetwork.Devnet;
+
+  // You can also provide a custom RPC endpoint
+  const endpoint = useMemo(() => clusterApiUrl(network), [network]);
+
+  // @solana/wallet-adapter-wallets includes all the adapters but supports tree shaking --
+  // Only the wallets you configure here will be compiled into your application
+  const wallets = useMemo(
+    () => [
+      new PhantomWalletAdapter(),
+      new GlowWalletAdapter(),
+      new SolletWalletAdapter({ network }),
+      new SolletExtensionWalletAdapter({ network }),
+    ],
+    [network]
+  );
+
+  return (
+    <ConnectionProvider endpoint={endpoint}>
+      <WalletProvider wallets={wallets} autoConnect>
+        {props.children}
+      </WalletProvider>
+    </ConnectionProvider>
+  );
+};